requirements.txt
agno
openai>=1.77.0
arize-phoenix
openinference-instrumentation-agno
opentelemetry-sdk
opentelemetry-exporter-otlp
atla-insights[openai]
# Pin discord.py-self to a commit known to work with Python 3.12
git+https://github.com/dolfies/discord.py-self.git@97e06c393f6e8f30d00fe0dd9cdf7196145fa851#egg=discord.py-self
e2b_code_interpreter
youtube_transcript_api
# Environment variable management
python-dotenv>=1.0.0
#

# Async HTTP support (may be needed for API calls)
aiohttp

serpapi
play-lichess
redis
groq
googlesearch-python
pycountry
wikipedia
exa-py
mcp
pydantic
uvicorn
pytz>=2024.1
asyncpg
